refactor(errors): one file per error
