+++ lib/errors/connection/access-denied-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const ConnectionError = require('./../connection-error');
+
+/**
+ * Thrown when a connection to a database is refused due to insufficient privileges
+ *
+ * @extends ConnectionError
+ */
+class AccessDeniedError extends ConnectionError {
+  constructor(parent) {
+    super(parent);
+    this.name = 'SequelizeAccessDeniedError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = AccessDeniedError;
