+++ lib/errors/connection/connection-acquire-timeout-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const ConnectionError = require('./../connection-error');
+
+/**
+ * Thrown when connection is not acquired due to timeout
+ *
+ * @extends ConnectionError
+ */
+class ConnectionAcquireTimeoutError extends ConnectionError {
+  constructor(parent) {
+    super(parent);
+    this.name = 'SequelizeConnectionAcquireTimeoutError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = ConnectionAcquireTimeoutError;
