+++ lib/errors/connection/connection-timed-out-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const ConnectionError = require('./../connection-error');
+
+/**
+ * Thrown when a connection to a database times out
+ *
+ * @extends ConnectionError
+ */
+class ConnectionTimedOutError extends ConnectionError {
+  constructor(parent) {
+    super(parent);
+    this.name = 'SequelizeConnectionTimedOutError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = ConnectionTimedOutError;
