+++ lib/errors/connection/host-not-found-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const ConnectionError = require('./../connection-error');
+
+/**
+ * Thrown when a connection to a database has a hostname that was not found
+ *
+ * @extends ConnectionError
+ */
+class HostNotFoundError extends ConnectionError {
+  constructor(parent) {
+    super(parent);
+    this.name = 'SequelizeHostNotFoundError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = HostNotFoundError;
