+++ lib/errors/database/exclusion-constraint-error.js
@@ -1,0 +1,26 @@
+'use strict';
+
+const DatabaseError = require('./../database-error');
+
+/**
+ * Thrown when an exclusion constraint is violated in the database
+ *
+ * @extends DatabaseError
+ */
+class ExclusionConstraintError extends DatabaseError {
+  constructor(options) {
+    options = options || {};
+    options.parent = options.parent || { sql: '' };
+
+    super(options.parent);
+    this.name = 'SequelizeExclusionConstraintError';
+
+    this.message = options.message || options.parent.message || '';
+    this.constraint = options.constraint;
+    this.fields = options.fields;
+    this.table = options.table;
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = ExclusionConstraintError;
