+++ lib/errors/database/timeout-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const DatabaseError = require('./../database-error');
+
+/**
+ * Thrown when a database query times out because of a deadlock
+ *
+ * @extends DatabaseError
+ */
+class TimeoutError extends DatabaseError {
+  constructor(parent) {
+    super(parent);
+    this.name = 'SequelizeTimeoutError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = TimeoutError;
