+++ lib/errors/query-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const BaseError = require('./base-error');
+
+/**
+ * Thrown when a query is passed invalid options (see message for details)
+ *
+ * @extends BaseError
+ */
+class QueryError extends BaseError {
+  constructor(message) {
+    super(message);
+    this.name = 'SequelizeQueryError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = QueryError;
