+++ lib/errors/sequelize-scope-error.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const BaseError = require('./base-error');
+
+/**
+ * Scope Error. Thrown when the sequelize cannot query the specified scope.
+ *
+ * @extends BaseError
+ */
+class SequelizeScopeError extends BaseError {
+  constructor(parent) {
+    super(parent);
+    this.name = 'SequelizeScopeError';
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
+
+module.exports = SequelizeScopeError;
